added voter routes
