+++ db/connection.js
@@ -1,0 +1,12 @@
+const mysql = require('mysql2');
+
+const db = mysql.createConnection({
+  host: 'localhost',
+  // Your MySQL username,
+  user: 'root',
+  // Your MySQL password
+  password: '',
+  database: 'election'
+});
+
+module.exports = db;
